Don't glob source files
This is a notorious CMake debate but I come down on the side of
strict correctness and avoid globbing. For a template project like
this where users are prone to be adding new source files regularly,
I worry that even with CONFIGURE_DEPENDS they may still have issues
with CMake not reconfiguring and not picking up the new .cpp file
they added. I'd prefer to leave globbing to experienced CMake users
who are likely not the audience of this template project.